requirements.txt
accelerate==1.10.0
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
defusedxml==0.7.1
filelock==3.19.1
fsspec==2025.7.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jira==3.10.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
protobuf==6.32.0
psutil==7.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
safetensors==0.6.2
sentencepiece==0.2.1
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
typing_extensions==4.14.1
urllib3==2.5.0
